feat(new-trace): Made bottom panel resizable. (#66453)
https://github.com/getsentry/sentry/assets/60121741/b4f469af-b88a-4a7d-b4de-480c1efcb93d

---------
